+++ components/adapters/pulumi/src/index.ts
@@ -1,0 +1,1 @@
+export { PulumiAdapter as default } from "./pulumi";