+++ components/adapters/pulumi/src/pulumi.ts
@@ -10,9 +10,9 @@ export class PulumiAdapter extends core.Adapter {
   private passphrase: string;
 
   //eslint-disable-next-line @typescript-eslint/no-var-requires
-  public readonly name = require(join(__dirname, "../../package.json")).name;
+  public readonly name = require(join(__dirname, "../package.json")).name;
   //eslint-disable-next-line @typescript-eslint/no-var-requires
-  public readonly version = require(join(__dirname, "../../package.json")).version;
+  public readonly version = require(join(__dirname, "../package.json")).version;
 
   constructor(args: core.NewAdapterArgs) {
     if (!isAbsolute(args.entrypoint)) {